requirements.txt
streamlit
google-generativeai
numpy
python-dotenv
